implements #4 in results GUI, various action fixes

+++ server/src/components/actions/FileDelete.ts
@@ -13,7 +13,7 @@ export default async function deleteFile(execute = false, act: Action, template:
         data.target = Handlebars.compile(action.target)(template);
         if (action.validate) if (!fs.existsSync(data.target)) return {warning: `Target path does not exist.`, data};
         if (!execute) return {data};
-        fs.unlinkSync(data.target);
+        if (fs.existsSync(data.target)) fs.unlinkSync(data.target);
         return { success: true, data};
     } catch (e){
         return {error: String(e), data};
