+++ server/src/components/actions/FolderDelete.ts
@@ -13,7 +13,7 @@ export default async function deleteFolder(execute = false, act: Action, templat
         data.target = Handlebars.compile(action.target)(template);
         if (action.validate) if (!fs.existsSync(data.target)) return {warning: `Target path does not exist.`, data};
         if (!execute) return {data};
-        fs.removeSync(data.target);
+        if (fs.existsSync(data.target)) fs.removeSync(data.target);
         return { success: true, data};
     } catch (e){
         return {error: String(e), data};
